Move util::time::duration_from_seconds to profile::time
