+++ components/profile/mem.rs
@@ -12,8 +12,8 @@ use std::borrow::ToOwned;
 use std::cmp::Ordering;
 use std::collections::HashMap;
 use std::thread;
+use time::duration_from_seconds;
 use util::thread::spawn_named;
-use util::time::duration_from_seconds;
 
 pub struct Profiler {
     /// The port through which messages are received.
